+++ tests/test_views.py
@@ -1336,7 +1336,7 @@ class TestBatchJobs:
             "returns": {"schema": {"type": "number"}},
         }
         if default_job_options is not None:
-            process_definition["default_job_parameters"] = default_job_options
+            process_definition["default_job_options"] = default_job_options
         requests_mock.get("https://share.test/add3.json", json=process_definition)
 
         pg = {