+++ openeo_driver/ProcessGraphDeserializer.py
@@ -58,7 +58,7 @@ from openeo_driver.errors import (
     CollectionNotFoundException, ProcessUnsupportedException,
 )
 from openeo_driver.processes import ProcessRegistry, ProcessSpec, DEFAULT_NAMESPACE, ProcessArgs
-from openeo_driver.processgraph import get_process_definition_from_url
+from openeo_driver.processgraph import get_process_definition_from_url, ProcessDefinition
 from openeo_driver.save_result import (
     JSONResult,
     SaveResult,
@@ -71,6 +71,7 @@ from openeo_driver.save_result import (
 from openeo_driver.specs import SPECS_ROOT, read_spec
 from openeo_driver.util.date_math import month_shift
 from openeo_driver.util.geometry import geojson_to_geometry, geojson_to_multipolygon, spatial_extent_union, BoundingBox
+from openeo_driver.util.http import is_http_url
 from openeo_driver.util.utm import auto_utm_epsg_for_geometry
 from openeo_driver.utils import EvalEnv, smart_bool
 
@@ -1813,7 +1814,7 @@ def apply_process(process_id: str, args: dict, namespace: Union[str, None], env:
         args = {name: convert_node(expr, env=env) for (name, expr) in sorted(args.items())}
 
     # when all arguments and dependencies are resolved, we can run the process
-    if namespace and any(namespace.startswith(p) for p in ["http://", "https://"]):
+    if is_http_url(namespace):
         if namespace.startswith("http://"):
             _log.warning(f"HTTP protocol for namespace based remote process definitions is discouraged: {namespace!r}")
         # TODO: security aspects: only allow for certain users, only allow whitelisted domains, support content hash verification ...?
@@ -2090,16 +2091,7 @@ def evaluate_process_from_url(process_id: str, namespace: str, args: dict, env: 
     :param process_id: process id of process that should be available at given URL (namespace)
     :param namespace: URL of process definition
     """
-    try:
-        process_definition = get_process_definition_from_url(process_id=process_id, url=namespace)
-    except OpenEOApiException:
-        raise
-    except Exception as e:
-        raise OpenEOApiException(
-            status_code=400,
-            code="ProcessNamespaceInvalid",
-            message=f"Process '{process_id}' specified with invalid namespace '{namespace}': {e!r}",
-        ) from e
+    process_definition: ProcessDefinition = get_process_definition_from_url(process_id=process_id, url=namespace)
 
     return _evaluate_process_graph_process(
         process_id=process_id,